refactor - moved form values validation functions to /utils, removed leftover comments
